backend/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
aspose-words==24.2.0
astunparse==1.6.3
blinker==1.7.0
certifi==2024.12.14
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==42.0.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
docs==0.1.0
docx==0.2.4
docx2pdf==0.1.8
docxlatex==0.1.6
environ==1.0
filelock==3.12.4
Flask==3.0.1
flatbuffers==24.12.23
fonttools==4.54.1
fpdf==1.7.2
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
looseversion==1.3.0
lxml==5.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mysqlclient==2.2.0
namex==0.0.8
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
opt_einsum==3.4.0
optree==0.13.1
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
platformdirs==3.11.0
protobuf==5.29.2
pyarrow==18.1.0
pycparser==2.21
Pygments==2.19.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PyMySQL==1.1.0
pyparsing==3.1.1
pypdfium2==4.26.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.66.2
traits==6.3.2
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.0
virtualenv==20.24.5
Werkzeug==3.0.1
wrapt==1.17.0
